Create grpc interceptor pkg, move auth interceptor into it
